pyproject.toml
[project]
name = "necro-net"
version = "0.1.0"
description = "Deep learning pathology image classification"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.9.4",
    "numpy>=1.24,<2.0",
    "onnx>=1.19.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "rich-argparse>=1.7.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch[onnx]>=2.2.0,<2.3",
    "torchvision>=0.17.0,<0.18",
    "tqdm>=4.67.1",
]
